requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
yfinance==0.2.31
pandas==2.0.3